+++ src/app/(layout)/header-zustand.tsx
@@ -1,12 +1,11 @@
 "use client";
 
 import { useLoginStore } from "@/hooks/zustand/store";
-import HomeHeader from "./home-header";
+import HeaderMobile from "./header-mobile";
 
 const HeaderWithZustand = () => {
   const loginStateKey = useLoginStore((state) => state.loginStateKey);
-
-  return <HomeHeader key={loginStateKey} />;
+  return <HeaderMobile key={loginStateKey} />;
 };
 
 export default HeaderWithZustand;
